+++ CHANGELOG.md
@@ -6,6 +6,26 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+
+- Added PR description generation
+
+  - Generate PR description based on the git diff output
+  - Create PRs with context-aware descriptions
+  - Allow users to refine the description until they are satisfied and ready to submit
+  - Allow users to specify a custom PR template
+
+- Support for Google GenAI provider (Google Gemini Flash)
+- Added Open Settings command to open extension settings faster
+
+### Changed
+
+- Added comments to all functions and classes in the codebase
+
+### Fixed
+
+- Skipped the lock file (e.g. package-lock.json) when retrieving the diff output
+
 ## [0.0.2] - 2024-10-29
 
 ### Changed